refactor!: revert to single-page layout
