--- app/src/routes/dashboard/(admin)/drafts/(ongoing)/regular/+layout.svelte
@@ -1,43 +1,0 @@
-<script>
-    import { AcademicCap, Beaker, PaperClip } from '@steeze-ui/heroicons';
-    import { TabAnchor, TabGroup } from '@skeletonlabs/skeleton';
-    import { Icon } from '@steeze-ui/svelte-icon';
-    import { page } from '$app/stores';
-</script>
-
-<!-- TODO: Ongoing Draft (ADMIN DASHBOARD) -->
-<TabGroup>
-    <!-- Tab Labels -->
-    <TabAnchor
-        href="/dashboard/drafts/regular/students"
-        selected={$page.url.pathname === '/dashboard/drafts/regular/students/'}
-        name="tab1"
-        value={0}
-    >
-        <Icon src={AcademicCap} slot="lead" class="h-8" />
-        <span>Registered Students</span>
-    </TabAnchor>
-    <TabAnchor
-        href="/dashboard/drafts/regular/labs"
-        selected={$page.url.pathname === '/dashboard/drafts/regular/labs/'}
-        name="tab2"
-        value={1}
-    >
-        <Icon src={Beaker} slot="lead" class="h-8" />
-        <span>Laboratories</span>
-    </TabAnchor>
-    <TabAnchor
-        href="/dashboard/drafts/regular/logs"
-        selected={$page.url.pathname === '/dashboard/drafts/regular/logs/'}
-        name="tab3"
-        value={2}
-    >
-        <Icon src={PaperClip} slot="lead" class="h-8" />
-        <span>System Logs</span>
-    </TabAnchor>
-
-    <!-- Tab Panels --->
-    <svelte:fragment slot="panel">
-        <slot />
-    </svelte:fragment>
-</TabGroup>
