--- app/src/routes/dashboard/(admin)/drafts/(ongoing)/regular/+page.js
@@ -1,5 +1,0 @@
-import { redirect } from '@sveltejs/kit';
-
-export async function load() {
-    redirect(302, '/dashboard/drafts/regular/students');
-}
