--- app/src/routes/dashboard/(admin)/drafts/noregistered/+page.svelte
@@ -1,8 +1,0 @@
-<script>
-    import WarningAlert from '$lib/alerts/Warning.svelte';
-</script>
-
-<WarningAlert
-    >No students have registered for this draft yet. This draft cannot proceed to the next round until at least one
-    student registers.</WarningAlert
->
